Move tests to dedicated folder
